patch-one from Ryan Schmidt
Attached is a patch that I believe brings over the useful changes from 2006-06-10 perl webdot while leaving out the typo and not removing the useful comment. I also moved the doubled slash fix and changed it slightly, and in the regex that removes the initial slash from the PATH_INFO, I made it clear that we want to remove the first slash only if it occurs at the beginning of the string. (That threw me off the search for the doubled slash issue for awhile.)