0002434: neatogen: fix spline not being created
We try to make a multi spline when there are multiple edges between two
nodes. We only make create this multi spline for the leader of these
edges.
If the leader already has a spline set and we are supposed to use given
splines (Nop > 1), then no multi spline will be created at all,
also not for other splines.

So we don't include these splines in the search for equivalent edges.