Check for unit test framework in configure.ac
Criterion is a unit testing framework for C and C++ and it's C99 compatible. It allows for fast test writing because, unlike a lot of other C unit test frameworks, it doesn't require a lot of boilerplate code.

This check in configure.ac checks if Criterion is installed on the system.