Tests added for current behaviour command line.
This commit adds some tests that "document" the current behavior of the command line interface. When refactoring command line interface (like the parsing code), the observable behavior of the command line interface should stay the same, this is to ensure backwards compatibility. These tests are a start, it covers some basic command line operations (printing the version and usage), but show some interesting things. Mainly that the order of the flags is very important.

- `dot -Vrandom` prints the version, but `dot -randomV` causes an error.
- `dot -?V` prints the usage, but `dot -V?` prints the version.

While it may be desirable to change this, when refactoring this behavior should stay the same. Changing the existing interface should be done in Graphviz 3.