Travis: Add build on CentOS 7 using Docker.
There are now two environments, one where the variable DOCKER_BUILD is true and one where it is false. False means a normal build on the Travis build image, true means docker will be used to run a build on a CentOS 7 image. The setup of the CentOS 7 image is done in the Dockerfile in `ci/centos7`. It takes a clean CentOS 7 image and installs the required build utilities and dependencies for a minimal build.