Removed old-style function defintions from lib/cdt
The functions definitions in cdt are all surrounded by an `#if __STD_C`, if __STD_C is not defined than the old style function definition is used. These days STD C is the standard and the old style function defintions are obsolete.
They have been removed from lib/cdt, but not from the subprojects.