Use `strcmp` in instead of `gvrender_comparestr`.
`bsearch` is used in this function to find a string in a string array. It's not needed to pass the address of `tok` to the function, because `tok` is already a pointer. `gvrender_comparestr` should then cast the values to `(const char)` instead of `*(char**)`, because `s1` and `s2` are strings, not pointers to a string. With this change the function only calls `strcmp`, so passing `strcmp` to `bsearch` has the same effect and `gvrender_comparestr` can be removed.
This gets rid of all the `-Wcast-qual` warnings as well.