Replace `SF_MAXDOUBLE` with `DBL_MAX` in lib/sfio
There are multiple checks to set SF_MAXDOUBLE, which is then only used
once. Since we are compiling with standard C and `float.h` is part of the
standard C API, we can use DBL_MAX instead and remove the define of
SF_MAXDOUBLE. This also results in some header files that are no longer
included, so checks for these could also be removed.