Remove all __DEFINE__ and __EXTERN__ macros
It seems that the condition where this macro is defined
(`#if !defined(_DLL_BLD) && defined(_dll_import)`) is never met. This
macro also seems to be meant for very specific scenario.
All in all, it seems safe to remove the macros by replacing them with
the value it gets when the above condition is not met.