Remove define and usage of `_hdr_float`
The header file float.h is part of the standard C library, so we can
assume it is always available. This header file also isn't included in the
sfio build when `_SFBINARY` is defined, resulting in a failed build.
The checks defining `_hdr_float` are removed and <float.h> is only
included were required in lib/sfio: in sfputd.c