Replace checks for <stdint.h> and <inttypes.h>
Instead of checking for the availability of stdint.h and inttypes.h with HAVE_STDINT_H and HAVE_INTTYPES_H, simply include stdint.h. The header stdint.h is part of the C99 standard, so it can be safely assumed that this header is present.
inttypes.h contains everything stdint.h includes, but in all cases here, stdint.h suffices.