Remove `_may_use_threads` from lib/sfio/vthread.h
The define `_may_use_threads` is set after checking some condition, but is
`undef`ed immediately after. The checks that follow only succeed if
`_may_use_threads` is defined, so they all fail and nothing happens. Therefore, the checks of this define and the usage of it can be removed.