Fixed: 2 'dereference before null check' in SparseMatrix
In the file lib/sparse/SparseMatric.x, there were two functions were a
pointer is dereferenced before checking if the pointer is NULL. The check
`if(!A)` caused an early return in both functions.
The fix is placing the check and the early return as the first line in
both functions. Dereferencing then only occures if the pointer A is not
NULL.