Prevent values like -0 in xdot output
Two graphs in the shapes regression test reference files (invhouse and
invtrapezium) contained a -0 coordinate. The Linux build produced the same
results, but the Windows build didn't, causing the regression tests to
fail.
To prevent values like -0 in the xdot output, a comparison is added just
before formatting the number.