Remove defintion of HAVE_STDARG_H and _hdr_stdarg
The define _hdr_stdarg was only used in ast_common.h to include <stdarg.h>
but this include was not needed, compilation succeeds without it.
The define HAVE_STDARG_H was only used to set _hdr_starg in ast_common.h.

Both defines and the include can therefore safely be removed.