Remove unused function checks and defines
congigure.ac checks for the presence of certain functions and creates
define indicating if it was found. Some of these defines are never
used, so the check (and the static define for Windows) can be safely
removed. This is the case for the following functions:
- __freadable
- _sysconf
- getrusage
- lsqrt
- memset
- nl_langinfo
- pow
- sqrt
- strchr
- strdup
- strstr
- strtoll
- strtoul
- strtoull
- uname

Also removed duplicate `strerror` in `AC_CHECK_FUNCS`.