Add install target and initial packaging to CMake
The installation locations are identical to the Autotools install step,
but the pdf version of the manpages aren't generated yet.
With CPack, packages can be generated based on the `install` commands. For
now only the basic ZIP package generator is added.
Added the new CMake commands to the Travis and Appveyor scripts.