Made xdot a dynamic library on Windows
xdot is a dynamic library with the Autotools build, but was for some
reason a static library on Windows. Since Autotools is considered the
default, this commit makes xdot a dynamic library on Windows as well.
The added define "XDOT_API", is used for linking, so a seperate xdot.def
file is not needed. On non-Windows platforms, this define is set to
"extern", so nothing changes there.