Remove usage of features/vmalloc
The only check left in features/vmalloc, checks if the return type of
malloc is the old `char *`. Because we are using standard C, we can
always assume the returntype of malloc is `void *`, this define is
there never set. We can remove this check and with it the usage of
features/vmalloc and FEATURE/vmalloc.