Merge Windows dependencies into one submodule
Instead of using a submodule for every Windows dependency with a lot of
extra files, we now use a single submodule that only contains those files
that are required. This means fewer submodules and shorter download time
(because of the reduction in size and number of files).
Also removed some unused links in Smyrna.