Made libgvpr a dynamic library on Windows
libgvpr is a dynamic library with the Autotools build, but was for some
reason a static library on Windows. Since Autotools is considered the
default, this commit makes libgvpr a dynamic library on Windows as well.
The added define "GVPR_API", is used for linking, so a seperate gvpr.def
file is not needed. On non-Windows platforms, this define is set to
"extern", so nothing changes there.