Add Windows installer using NSIS and CPack
Create a Windows Installer using CPack and the NSIS generator. The CMake
build on Appveyor now installs Graphviz using this installer and runs the
tests using ctest. The installer is added as an artifact to the Release
build.
Also switched the order of cmake and msbuild in Appveyor and added
`clone_depth: 25` for faster cloning.