Reduce exported symbols for plugin/neato_layout
On Windows, the plugin gvplugin_neato_layout exported a lot of function,
also from other libraries that are not being exported, such as ortho and
neatogen. This has been reduced to only the varaible
`gvplugin_neato_layout_LTX_library`.
A minor link change was needed on Windows as a result.