Add lib/ast to include directories for lib/gvc
The addition of including `unistd.h` or `compat_unistd.h` in gvplugin.c
caused the build to fail on Windows because `compat_unistd.h` could not be
found. The compatibility header is located in `lib/ast`, which is now
added to the include directories. This resolves the error of the failed
build on Windows.